requirements.txt
pip>=9
setuptools>=26
wheel>=0.29
pytest
coverage
flake8
black
yapf
python-gitlab
twine
six==1.15.0
numpy
joblib
memoized-property
mlflow
s3fs
termcolor
gcsfs==2021.5.0
google-cloud-storage==1.38.0
pandas==1.2.3
pandas-profiling==2.11.0
scikit-image==0.18.1
scikit-learn==0.22
uvicorn==0.13.4
pytz
fastapi
memoized_property

